refactor inputs
