+++ src/pages/scientificServices/pipelines/tabs/run/inputs/boolean/PipelineBooleanInput.test.tsx
