+++ src/pages/scientificServices/pipelines/tabs/run/inputs/float/PipelineFloatInput.test.tsx
