+++ src/pages/scientificServices/pipelines/tabs/run/inputs/string/PipelineStringInput.test.tsx
