+++ src/pages/scientificServices/pipelines/tabs/run/inputs/string/PipelineStringInput.tsx
